windows/requirements.txt
pywebview[edgehtml]>=5.0.0
cryptography==41.0.5
colorama==0.4.6
cheroot==10.0.1
cryptography==41.0.5